Started Chapter 6 example KeyboardAndFocusDemo. Still need to complete paintComponent().
Am now using Notepad++ without syntax highlighting or any other kind of 'hints' for majority of development, using NetBeans only when setting up a new package/project, in order to prepare for the exam where we will have to code without any hints. In other words, it's going to be slower, with more errors, and probably not as good. Don't worry NetBeans, I'll be back...